check: Add SFINAE checks for well formed thermal materials
